refactor: move search+item list+pagination to the separate component
